requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pymongo==4.6.0

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Message Queue
kafka-python==2.0.2
celery==5.3.4

# Data Sources
pyodbc==5.0.1  # MSSQL
happybase==1.2.0  # HBase
gemfire-python-client==10.2.0  # Gemfire (if available)

# Monitoring & Metrics
prometheus-client==0.19.0
structlog==23.2.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Alerting
smtplib-ssl==1.0.0
twilio==8.11.0

# Dashboard
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Utilities
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
